+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [Unreleased]
+### Added
+* [#394](https://github.com/shlinkio/shlink-frontend-kit/issues/394) Add first tailwind-based components
+* [#298](https://github.com/shlinkio/shlink-frontend-kit/issues/298) Add `@shlinkio/shlink-frontend-kit/tailwind` entry point to load tailwind-based components
+
+### Changed
+* *Nothing*
+
+### Deprecated
+* *Nothing*
+
+### Removed
+* *Nothing*
+
+### Fixed
+* *Nothing*
+
+
 ## [0.7.3] - 2025-03-07
 ### Added
 * Add new `determineOrder` helper function, which works like `determineOrderDir`, but returns the full `Order` object, rather than just the new dir.